Re-organized requirements files
The base requirements no longer make sense given the variety of scripts with individual requirements. Additionally, the Sailthru scripts have been moved to a parent package.
