+++ auditing/requirements.txt
@@ -1,0 +1,2 @@
+MySQL-python==1.2.5
+pandas==0.16.2
