+++ google_analytics/ga_homepage_report.py
@@ -1,12 +1,12 @@
 import argparse
-from datetime import datetime, timedelta
-import httplib2
 import sys
+from datetime import datetime, timedelta
 
+import httplib2
 from apiclient.discovery import build
 from oauth2client import client, file, tools
 from pandas import DataFrame, ExcelWriter, merge, concat, to_numeric
-import xlsxwriter
+
 
 ################################################################
 #
