--- requirements.txt
@@ -1,2 +1,0 @@
--r requirements/base.txt
-
