--- requirements/base.txt
@@ -1,8 +1,0 @@
-edx-rest-api-client==1.6.0
-google-api-python-client==1.5.3
-MySQL-python==1.2.5
-numpy==1.11.1
-pandas==0.16.2
-sailthru-client==2.2.3
-vertica-python==0.6.7
-XlsxWriter==0.9.3
