+++ sailthru/requirements/base.txt
@@ -1,0 +1,2 @@
+edx-rest-api-client==1.6.0
+sailthru-client==2.2.3
